messages, invitations seem fully working

+++ app/models/invitation.rb
@@ -3,16 +3,28 @@ class Invitation < Message
   belongs_to :invitable, :polymorphic => true
   validates_presence_of :invitable_id
   polymorphic_attr :invitable, :as => [ :event, :group, :contact ], :class_names => { :contact => 'User' }
+  validate :no_preexisting_relationship
   
 
   def after_accepted
     if event?
-      event.rsvps.create :user => recipient
+      new_item = recipient.rsvps.find_or_create_by_event_id( event_id )
     end
     if group?
-      group.memberships.create :user => recipient
+      new_item = recipient.memberships.find_or_create_by_group_id( group_id )
     end
-    self.destroy
+    if contact?
+      new_item = recipient.contact_relationships.find_or_create_by_contact_id( contact_id )
+    end
+    self.destroy if new_item && new_item.save
+  end
+
+  def no_preexisting_relationship
+    unless recipient.nil?
+      errors.add( :event, "already being attended by this member" ) if event? and recipient.events.include?( event )
+      errors.add( :group, "already includes this member" ) if group? and recipient.groups.include?( group )
+      errors.add( :recipients, "are already in your contacts" ) if contact? and recipient.contacts.include?( contact )
+    end
   end
 
 end
